Store trajectory duration as f64 instead of truncating to u32
get_duration computes totaltime_ns as a float, but Duration.totaltime_ns
and ExportSimulation.duration were u32, so the value was cast down before
export. Some trajectories (such as the DD-13M dissociation data) are
sub-nanosecond (e.g. 0.44 ns), and were floored to 0 in the
database.

The DB column (duration) is already Nullable<Float8> and the mdr-db model
is Option<f64>, so widening these two fields to f64 and dropping the cast
sends the real value through with no migration. Verified with a dry-run:
import.json now reports "duration": 0.44 where it previously emitted 0.

(code implemented with help from Claude Opus 4.8)

+++ mdr-process/src/types.rs
@@ -273,7 +273,7 @@ pub struct RmsdRmsf {
 #[derive(Debug, Deserialize, Serialize)]
 #[serde(deny_unknown_fields)]
 pub struct Duration {
-    pub totaltime_ns: u32,
+    pub totaltime_ns: f64,
     pub sampling_frequency_ns: f32,
 }
 
@@ -444,7 +444,7 @@ pub struct ExportSimulation {
 
     pub rmsf_values: Vec<f64>,
 
-    pub duration: u32,
+    pub duration: f64,
 
     pub sampling_frequency: f32,
